fixed warnings except the one related to auto_ptr deprecated in unit tests, we'll have to wait for a more recent cppunit lib

+++ src/libws_diff/DiffPrinter/jsonDiffPrinter.cc
@@ -31,7 +31,7 @@ void jsonDiffPrinter::addInfo(const std::string& key,const double value){
 void jsonDiffPrinter::addRequestUri(const std::string& uri,
 		const std::string& paramsBody){
 	this->jsonRes["request"]["uri"]=uri;
-	int pivot=uri.find('?');
+	size_t pivot=uri.find('?');
 	this->jsonRes["request"]["url"]=uri.substr(0,pivot);
 	std::vector<std::string> vectParams;
 	if(pivot!=uri.length()){
